backend/requirements.txt
falcon>=3.0,<4.0
uvicorn
celery
asyncpg
uvicorn
redis
python-dotenv